requirements.txt
aotools==1.0.7
astropy==6.1.1
baldr==1.0
bmc==0.0.4
control==0.10.1
corner==2.2.2
fpdf==1.7.2
matplotlib==3.8.3
numpy==1.23.0
opencv_python==4.10.0.84
pandas==2.2.3
parse==1.19.0
pillow==10.2.0
PyQt5==5.15.10
PyQt5_sip==12.13.0
pyqtgraph==0.13.7
pyserial==3.5
PySpin==1.1.1
PyVISA==1.13.0
pyvisa-py==0.6.3
pyzelda==1.2
psutil==5.9.0
scikit_learn==1.6.1
scipy==1.9.1
setuptools==68.2.2
scikit-image==0.20.0
streamlit==1.39.0
sympy==1.13.2
tqdm==4.65.0
zaber_motion==6.5.0
zmq==0.0.0
zeroconf==0.131.0